requirements.txt
numpy
Pillow
matplotlib
requests
